Implemented GIT functionality to card_set
Now card banks are managed by git
The process is quite simplified in comparison to how much stuff git usually can provide, that is by design, to ensure the tool is not too overwhelming.

I am having a problem related to deleting repositories and uninstalling the app. That has to do with how Python deals with read only files in windows, Though that is not crucial for my tool to work, its annoying users will need to navigate to the folder system to delete the files